requirements.txt
spotipy
click
vyper-config
Pillow>=10.1
inky[rpi,example-depends]